Rearchitect to be closer to original code
