+++ CavernSamples/CavernizeGUI/MainWindow.Shutdown.cs
@@ -1,0 +1,6 @@
+namespace CavernizeGUI;
+
+partial class MainWindow {
+    static void CheckBlocks() {
+    }
+}
